+++ builtin/providers/aws/resource_aws_vpc_test.go
@@ -30,6 +30,35 @@ func TestAccVpc(t *testing.T) {
 	})
 }
 
+func TestAccVpcUpdate(t *testing.T) {
+	var vpc ec2.VPC
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckVpcDestroy,
+		Steps: []resource.TestStep{
+			resource.TestStep{
+				Config: testAccVpcConfig,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckVpcExists("aws_vpc.foo", &vpc),
+					testAccCheckVpcCidr(&vpc, "10.1.0.0/16"),
+					resource.TestCheckResourceAttr(
+						"aws_vpc.foo", "cidr_block", "10.1.0.0/16"),
+				),
+			},
+			resource.TestStep{
+				Config: testAccVpcConfigUpdate,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckVpcExists("aws_vpc.foo", &vpc),
+					resource.TestCheckResourceAttr(
+						"aws_vpc.foo", "enable_dns_hostnames", "true"),
+				),
+			},
+		},
+	})
+}
+
 func testAccCheckVpcDestroy(s *terraform.State) error {
 	conn := testAccProvider.ec2conn
 
@@ -100,5 +129,12 @@ func testAccCheckVpcExists(n string, vpc *ec2.VPC) resource.TestCheckFunc {
 const testAccVpcConfig = `
 resource "aws_vpc" "foo" {
 	cidr_block = "10.1.0.0/16"
+}
+`
+
+const testAccVpcConfigUpdate = `
+resource "aws_vpc" "foo" {
+	cidr_block = "10.1.0.0/16"
+	enable_dns_hostnames = true
 }
 `